requirements.txt
jinja2
requests
click
pwntools
tqdm
python-magic
beautifulsoup4
lxml
prettytable